renamed installGlassfish (#6761)
