+++ scripts/installer/installAppServer.py
